RestTokenTest: set queryTime=true at init
RestTokenTest.authtime2 was setting queryTime=true for its requestToken,
which doesn't do what it thinks because queryTime is only used at
library init. Also the way it was doing it will not work after the 0.9
change where supplying a non-null AuthOptions overrides rather than
merges with the stored defaults.

Fixed.