0.9 TokenParams and AuthOptions handling
Auth.requestToken, Auth.createTokenRequest and Auth.authorize all use
default TokenParams and AuthOptions only if null is passed as the call
arg. If null is not passed for either call arg, the supplied
TokenParams/AuthOptions (respectively) is used in its entirety; there is
no merge with the default configured values.
(RSA8e for requestToken, RSA9h for createTokenRequest, RSA10j for
authorise)

This is an incompatible change, so should not be merged until 0.9.